Move pure types back to shared types
