+++ src/options.ts
@@ -1,4 +1,4 @@
-import type { AutoDict, HandlerOptions } from "./pure_types";
+import type { AutoDict, HandlerOptions } from "./shared_types";
 import type { Controller } from "./services/textarea";
 
 export type SubstituteKeyboardEvents = (
